Bug fix for very long author list parsing error
Turns out that the feed truncates the authors list at some
prededfined character length, and then parses the authors.  This
means that the last author listed can have a garbled name and that
there can be some authors missing.  This appears to be a problem
in the way that bioRxiv is implementing the feed or the way I'm
requesting it.

Either way, this fix just ensures all authors have a comma in
their names => a last name and (partial) first name exists.  I'll
think about a fix for making sure all authors are parsed (this is
important for matching authors in the conf file who are often last
authors or PIs.